Remove daemon support
The "daemon" feature contains a security vulnerability [1].

This feature relies on a beta release (e.g. version 0.5.1) of a Node.js
module on npm--one that was superseded by a stable (e.g. version 1.0)
release published three years ago [2]. Due to a build-time dependency on
the long-since deprecated `node-waf` tool, the module at that version
can only be built for Node.js versions 0.8 and below.

Given this, actual usage of this feature is likely very limited. Remove
it completely so the integrity of this module's core functionality can
be verified.

[1] https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=809252
[2] The source expects the module to define a property named
    `deamonize`, which was removed for the 1.0 release:
    https://github.com/indexzero/daemon.node/tree/v1.0.0